refactor: refactor all files
