+++ src/components/footer/footer.module.css
@@ -11,10 +11,11 @@
 
 hr {
   display: flex;
-  width: 100%;
+  width: 89vw;
 }
 
 .icon {
+  bottom: -20%;
   display: flex;
   justify-content: center;
   gap: 25px;
@@ -23,7 +24,7 @@ hr {
 
 .description {
   text-align: center;
-  margin: 0 auto;
+  margin: 55px auto 0;
 }
 
 .description p {
@@ -36,4 +37,23 @@ hr {
 
 .description a {
   color: #EF4934;
+}
+
+@media only screen and (min-width: 1250px) {
+
+  .footer {
+    position: static;
+    justify-content: center;
+    padding: 0 30px;
+    gap: 30px;
+    height: 166px;
+    align-items: center;
+    flex-shrink: 0;
+  }
+
+  .hr {
+    display: flex;
+    width: 700px;
+    justify-content: space-between;
+  }
 }
